--- a/benches/lib_benches.rs
+++ b/benches/lib_benches.rs
@@ -22,6 +22,27 @@ macro_rules! bench_decimal_op {
     }
 }
 
+macro_rules! bench_fold_op {
+    ($name:ident, $op:tt, $init:expr) => {
+        #[bench]
+        fn $name(b: &mut ::test::Bencher) {
+            fn fold(values: &[Decimal]) -> Decimal {
+                let mut acc: Decimal = $init.into();
+                for value in values {
+                    acc = acc $op value;
+                }
+                acc
+            }
+
+            let values: Vec<Decimal> = test::black_box((0..10_000).map(|i| i.into()).collect());
+            b.iter(|| {
+                let result = fold(&values);
+                ::test::black_box(result);
+            });
+        }
+    }
+}
+
 /* Add */
 bench_decimal_op!(add_one, +, "1");
 bench_decimal_op!(add_two, +, "2");
@@ -31,23 +52,7 @@ bench_decimal_op!(add_negative_point_five, +, "-0.5");
 bench_decimal_op!(add_pi, +, "3.1415926535897932384626433832");
 bench_decimal_op!(add_negative_pi, +, "-3.1415926535897932384626433832");
 
-#[bench]
-fn bench_sum_10k(b: &mut ::test::Bencher) {
-    fn sum_10k(values: &[Decimal]) -> Decimal {
-        let mut sum: Decimal = 0.into();
-        for value in values {
-            sum = sum + value;
-        }
-        sum
-    }
-
-    let values: Vec<Decimal> = test::black_box((0..10_000).map(|i| i.into()).collect());
-    b.iter(|| {
-        let result = sum_10k(&values);
-        ::test::black_box(result);
-    });
-}
-
+bench_fold_op!(sum_10k, +, 0);
 
 /* Sub */
 bench_decimal_op!(sub_one, -, "1");
@@ -58,22 +63,7 @@ bench_decimal_op!(sub_negative_point_five, -, "-0.5");
 bench_decimal_op!(sub_pi, -, "3.1415926535897932384626433832");
 bench_decimal_op!(sub_negative_pi, -, "-3.1415926535897932384626433832");
 
-#[bench]
-fn bench_dec_10k(b: &mut ::test::Bencher) {
-    fn dec_10k(values: &[Decimal]) -> Decimal {
-        let mut sum: Decimal = 5_000_000.into();
-        for value in values {
-            sum = sum - value;
-        }
-        sum
-    }
-
-    let values: Vec<Decimal> = test::black_box((0..10_000).map(|i| i.into()).collect());
-    b.iter(|| {
-        let result = dec_10k(&values);
-        ::test::black_box(result);
-    });
-}
+bench_fold_op!(dec_10k, -, 5_000_000);
 
 /* Mul */
 bench_decimal_op!(mul_one, *, "1");