This changes simplify `add` function and make big speed up via agressive inlining functions of used for addition `Decimal`.
Actually, after adding `bench_dec_10k`, `bench_sum_10k` became show almost 1.75x worst result. Adding attribute `#[inline(always)]` for `add` and `sub` fixed it. I hope it have such positive effect in common, but I'm not sure about it.

Benchmarks:

before
```
test add_negative_pi         ... bench:         171 ns/iter (+/- 6)
test add_negative_point_five ... bench:          28 ns/iter (+/- 4)
test add_one                 ... bench:          29 ns/iter (+/- 4)
test add_one_hundred         ... bench:          29 ns/iter (+/- 1)
test add_pi                  ... bench:         166 ns/iter (+/- 11)
test add_point_zero_one      ... bench:          14 ns/iter (+/- 3)
test add_two                 ... bench:          29 ns/iter (+/- 4)
```
after
```
test add_negative_pi         ... bench:         164 ns/iter (+/- 4)
test add_negative_point_five ... bench:          21 ns/iter (+/- 3)
test add_one                 ... bench:          26 ns/iter (+/- 1)
test add_one_hundred         ... bench:          26 ns/iter (+/- 3)
test add_pi                  ... bench:         164 ns/iter (+/- 22)
test add_point_zero_one      ... bench:          11 ns/iter (+/- 0)
test add_two                 ... bench:          26 ns/iter (+/- 4)
```

before
```
test sub_negative_pi         ... bench:         168 ns/iter (+/- 16)
test sub_negative_point_five ... bench:          24 ns/iter (+/- 4)
test sub_one                 ... bench:          35 ns/iter (+/- 2)
test sub_one_hundred         ... bench:          36 ns/iter (+/- 4)
test sub_pi                  ... bench:         173 ns/iter (+/- 12)
test sub_point_zero_one      ... bench:          22 ns/iter (+/- 3)
test sub_two                 ... bench:          35 ns/iter (+/- 2)
```
after
```
test sub_negative_pi         ... bench:         164 ns/iter (+/- 13)
test sub_negative_point_five ... bench:          21 ns/iter (+/- 1)
test sub_one                 ... bench:          27 ns/iter (+/- 0)
test sub_one_hundred         ... bench:          28 ns/iter (+/- 0)
test sub_pi                  ... bench:         164 ns/iter (+/- 13)
test sub_point_zero_one      ... bench:          12 ns/iter (+/- 0)
test sub_two                 ... bench:          27 ns/iter (+/- 2)
```

before
```
test bench_sum_10k           ... bench:      79,669 ns/iter (+/- 5,988)
```
after
```
test bench_sum_10k           ... bench:      41,334 ns/iter (+/- 1,217)
```

before
```
test bench_dec_10k           ... bench:     109,814 ns/iter (+/- 5,145)
```
after
```
test bench_dec_10k           ... bench:      46,768 ns/iter (+/- 2,222)
```

C# version .net core
```
     Method |     Mean |     Error |    StdDev |
----------- |---------:|----------:|----------:|
 DecDecimal | 109.2 us | 0.6073 us | 0.5384 us |
 SumDecimal | 112.0 us | 0.0913 us | 0.0854 us |
```